stories for the selection screen and arena
The ones labeled just "story 1" or "story 2" are meant for the story selection screen. Then the others with "green" or "red" are for the arena scene. "Green" is the regular state of increasing completion of the story and "red" is the state of decreasing the completion amount when a bug reaches the bottom of the screen.